refactor: refactor some code to make it more readable
1. modify go.mod module name.
2. move disk.go and mounter.go into nvmf.go
3. make Connector as an objective
