+++ go.mod
@@ -1,15 +1,16 @@
-module csi-driver-nvmf
+module github.com/kubernetes-csi/csi-driver-nvmf
 
 go 1.14
 
 require (
 	github.com/container-storage-interface/spec v1.5.0
-	github.com/golang/glog v0.0.0-20210429001901-424d2337a529
-	github.com/kubernetes-csi/csi-lib-utils v0.10.0
+	github.com/golang/glog v1.0.0
+	github.com/google/go-cmp v0.5.6 // indirect
+	github.com/kubernetes-csi/csi-lib-utils v0.11.0
 	github.com/sirupsen/logrus v1.8.1
-	github.com/spf13/cobra v1.2.1
-	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d
-	google.golang.org/grpc v1.39.1
+	golang.org/x/net v0.0.0-20220225172249-27dd8689420f
+	google.golang.org/genproto v0.0.0-20211208223120-3a66f561d7aa // indirect
+	google.golang.org/grpc v1.44.0
 	k8s.io/klog v1.0.0
-	k8s.io/utils v0.0.0-20210802155522-efc7438f0176
+	k8s.io/utils v0.0.0-20220210201930-3a6ce19ff2f9
 )
